Refactor graphics.
