+++ src/ansys/fluent/visualization/graphic/__init__.py
@@ -2,5 +2,5 @@
 
 from ansys.fluent.visualization.graphic.graphic_objects import Graphics  # noqa: F401
 from ansys.fluent.visualization.graphic.graphic_windows_manager import (  # noqa: F401
-    graphic_windows_manager,
+    graphics_windows_manager,
 )
